requirements.txt
azure-ai-textanalytics
azure-ai-translation-text
azure-core
black
pylint
python-dotenv
pytest